# lib/api-zod/src/generated/types/adminAuthData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Affiliate Marketing Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUser } from "./adminUser";

export interface AdminAuthData {
  success: boolean;
  user: AdminUser;
}