# lib/api-zod/src/generated/types/adminLoginInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Affiliate Marketing Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginInput {
  username: string;
  password: string;
}