# lib/api-zod/src/generated/types/clicksTimePoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Affiliate Marketing Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ClicksTimePoint {
  date: string;
  clicks: number;
}