# lib/api-zod/src/generated/types/offer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Affiliate Marketing Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";

export interface Offer {
  id: number;
  title: string;
  /** @nullable */
  titleAr?: string | null;
  slug: string;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  shortDescriptionAr?: string | null;
  /** @nullable */
  longDescription?: string | null;
  /** @nullable */
  longDescriptionAr?: string | null;
  affiliateUrl: string;
  imageUrl: string;
  /** @nullable */
  ctaText?: string | null;
  /** @nullable */
  ctaTextAr?: string | null;
  rating: number;
  /** @nullable */
  categoryId?: number | null;
  category?: Category | null;
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoTitleAr?: string | null;
  /** @nullable */
  seoDescription?: string | null;
  /** @nullable */
  seoDescriptionAr?: string | null;
  /** @nullable */
  keywords?: string | null;
  /** @nullable */
  faqSchema?: string | null;
  isActive: boolean;
  isFeatured: boolean;
  isTrending: boolean;
  sortOrder?: number;
  clickCount: number;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}