# lib/api-zod/src/generated/types/offerClickStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Affiliate Marketing Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface OfferClickStat {
  offerId: number;
  title: string;
  slug: string;
  imageUrl: string;
  totalClicks: number;
}