# lib/api-zod/src/generated/types/offerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Affiliate Marketing Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface OfferInput {
  /** @minLength 1 */
  title: string;
  titleAr?: string;
  /** @minLength 1 */
  slug: string;
  shortDescription?: string;
  shortDescriptionAr?: string;
  longDescription?: string;
  longDescriptionAr?: string;
  affiliateUrl: string;
  imageUrl: string;
  ctaText?: string;
  ctaTextAr?: string;
  /**
   * @minimum 0
   * @maximum 5
   */
  rating?: number;
  categoryId?: number;
  seoTitle?: string;
  seoTitleAr?: string;
  seoDescription?: string;
  seoDescriptionAr?: string;
  keywords?: string;
  faqSchema?: string;
  isActive?: boolean;
  isFeatured?: boolean;
  isTrending?: boolean;
  sortOrder?: number;
}